Rework things to remove a lot of allocations
Rather than use a heirarchy of structs with a pointer the object
stored in the base object.  In addition to removing a lot of small
allocations, this means that we get the same address each time for a
Go object that represents a specific Python object (whereas before we
would get a new Go object each time).