If we call actual on a nil pointer, return nil
This is actually a useful thing to do, as otherwise we end up with a
nil pointer inside the interface, instead of a nil interface value.
This is a subtle distinction, but means "v == nil" will return false
if v is an Object with a nil pointer in instead of a nil interface
value.