Improve the way exception() works
For one thing, it now detects if an exception has not been raised and
returns nil, otherwise it will always return an Error (even if the
value is nil, which it may be).

It's also tidier code, and returns an Error (including the tb, should
code want to use it) instead of using fmt.Errorf.