Check function signatures of class methods
When the user defines methods on a struct to implement methods for the
Python class then we need the function signature to be correct - since
we basically bypassing the checking that Go would normally do for us.

If the user writes a method with the wrong signature, and we call it
without checking the code will panic when (if) that function gets
called in a very unobvious manner.

By checking the signatures when we create the class we can spit out a
helpful error nice an early - and the problem shouldn't every get past
the first run.