Start adding support for exporting fields to Python
If a field in a struct that implements a Python class has either a Py
or PyDoc tag, then we export that field to Python (using the Py tag
value as the name if specified).  This is equivalent to the tp_members
of a type object in the C API.

Currently only objects (i.e. things that implement the Object
interface, or Object fields) are supported, native types are
identified by not yet exported.