Start implementing the number protocol
The number protocol is rather large - this is only the beginnings of
implementing it.  This gets us as far as calling a Go function if the
appropriate method has been defined.  The missing bit is actually
turning the function call into the method call, the rest should be in
place now.