Complete the number protocol implementation
By using helper functions it was actually quite easy to implement all
the methods of the number protocol.

Well, all except coerce, but that isn't needed for newer style usage
where the functions do their own type worrying.  This is the only
mechanism supported by Python 3, and the only one we support (for now
anyway).