Add a convienience Alloc method for Go Object creation
When creating an Object instance of a Go defined type in Go, we need
to provide an easy way for the user to actually allocate the object.
Since they may not want to use their PyNew method (or they might not
have written one), we provide an Alloc method on Class that not only
allocates a new Object, but also registers it so that we will be able
to find the context later.