Finally implement PackTuple
Finally got around to implementing PackTuple.  This function
replicates the functionality of the C function rather than calling it.
This is because the function is a) fairly trivial, b) rather hard to
call from Go.  Rather than adding a lot of overhead to the call, we
simply do the work in Go ourselves.  Though we do have to do some
messy pointer manipulation as Go won't let us use indexing.