Make it possible to use a channel from Python
Channels are useful things for communicating.  Make it possible to
communicate with Python code by wrapping a "chan Object" into a Chan,
an object that can be used in Python.

The Python object has put and get methods (modelled after the Queue
type) that allow Python to send and recieve on the underlying channel.

We also introduce a "go" module that contains the Chan type from the
Python side.  The "go" module must be initialized before the new Chan
type can be used.