Start splitting BaseObject into two parts
Until now we have been using BaseObject for two purposes.
Unfortunately this doesn't work when we want to extend Python's base
types in Go code.

AbstractObject takes over the role of the type that provides the
methods to satisfy the Object interface via embedding.  AbstractObject
has the advantage of being of 0 size, so it can be embedded in the
specific types whilst leaving them able to have the correct size.
This is important as it allows the types such as List, Dict etc. to
have the correct size - allowing them to be extended by the user (or
at least, it will when the class code has support for subclassing).

BaseObject is now the Go version of PyObject - in particular it can be
used when implementing types in Go.  It is also gives us a type to
implement the PyObject_XXX methods that we don't want to make part of
the Object interface.