Separate AbstractObject and BaseObject better
Now that we have two types, we can a) put them in separate files for
mental separation, b) have AbstractObject only implement the common
methods in the Object interface.  We actually split object.go into
three files, abstract.go implements the methods on the empty
AbstractObject, and base.go implements the methods on the concrete
BaseObject, and now object.go is somewhere to put the miscellaneous
object related bits (which were already there, but largely hidden by
the other stuff).