Make it possible to initialise an arbitary type
When subclassing a native type, we need to be able to call the init
function for the type we have subclassed.  To make this possible we
add an Init method to the Type type.  We also need to expose the Type
objects for each of the native types - though we have only add
ListType so far.