We don't need the setBase function
The setBase function was intended to make it easy to duplicate the
messy setting of tp_base.  However, we are using a map instead of the
original switch so don't need to duplicate the setting - which can
also be simpler anyway, if we cast to the correct type (it's struct
_typeobject, not struct typeobject - I was missing an _ in previous
attempts).