Add convience function for Init when using threads
There is an interesting issue with using Python from multiple
goroutines.  Since Go multiplexes goroutines onto threads, we don't
know which thread we are running on.  The Lock type is supposed to
sort all this out for us - but it does require that Python be
initialised in the correct way to work reliably.  Add a convienience
function that does the initialisation and thread setup in a way that
leaves things in a state where Locks work as expected.