Add a second, more useful, example
The new example is a very simple shell.  This shell is written in Go,
but allows commands to be implemented in Python - and gives them
access to tokenisation and run commands via a sh module.

This demonstrates calling into Python, and extending Python with Go
modules.  It doesn't include creating a new type in Go though.