Improve InitAndLock
The InitAndLock function is supposed to be the safe way to initialize
Python.  However, we need to lock the current goroutine to the current
thread whilst we initialize Python - as the interpreter uses
per-thread state, and we want that state to remain the same until we
can release the GIL ready for any code to grab the GIL and make its
own thread state current.  Once this is done, then we can create and
lock a new Lock.