Improve handling of Exceptions
Still not as nice as I would like, but this is closer.

Unfortunately, Python doesn't provide a way to get the message that
would be displayed to the user as a string - it can only be printed to
stderr.  So, we have to provide our own mechanism constructed from the
primatives that _are_ available.  This makes the String method a
little unpleasant.