Pass proper errors to Python
In places where code had to raise an exception in Python we were using
fmt.Errorf to create an error that we turned into a generic exception.
This isn't appropriate - and we now have a way to do things
differently.  So change all the uses of Errorf to create an exception
for Python to create specific types of exception.  Also fix some
places where we using Printf instead of raise an exception at all.