Go classes need to be allocated in Go
Because Go is a GC language, we need to allocate the memory that we
use for classes implemented in Go in a way that makes that memory
visible to the Go Garbage Collector.  If we don't do this then Go may
decide that it can free some memory that we are actually still using.

This problem tends not to show up in simple tests, as they simply
don't use enough memory for the Go GC to kick in.  Which is why I
hadn't noticed this until recently.

Unfortunately, neither Go nor Python make it particularly easy to play
around with memory allocation and GC tracking.  So the new memory.go
is rather icky.  Not only having to replicate some Python code, but
also trick both the Python and Go GC code ... :(