Start locking global state
There are some global variables that are used to maintain global
state.  These variables need to be protected if we are going to be
accessing them from multiple goroutines (which we may well be).

Not all of the variables are locked by this change.  Hopefully those
that aren't have at least got a TODO now though ...