Change some locks to RWMutex
Some of the locks recently introduced make more sense as RWMutex than
as Mutex.  They are all protecting state that is likely to be accessed
a lot more than it is modified.  There is no need to serialise
accesses, only to make sure that there is no access in parallel with a
modify - which is exactly the protection a RWMutex gives us.