Try and generate gccgo version from gc version
It would be nice to be able to keep the gccgo version in step with the
gc version.  This is an attempt to automate the necessary changes.  At
the moment only the Go code is being processed, but eventually the C
code will need to be done too (which will probably involve splitting
out the gccgo specific code from that shared with the gc
implementation).