Move VC only files to separate make file
There are some commands that need to be inside VC, and some rules that
will break things if included outside of VC.  So move a bunch of stuff
from the Makefile to a separate file that isn't included in the
tarball.  The Makefile can then -include this file, so it will be
ignored if not available.