Make it possible to call Go from threads
Previously it was not possible to call into Go from anything other
than the thread that initialised the Go runtime.  We still can't run
Go code on threads created outside of the Go runtime - but we can
create proxy threads inside the runtime.

We start a dispatcher goroutine when we start up the runtime, and then
when an attempt to call into Go is made on another thread we instead
send a message to this goroutine then then creates a proxy goroutine
that runs the function for us.  This goroutine runs locked to a thread
and can call back into Python (though in it's thread not the one of
the original call).

The only potential issue (ignoring performance) with this that I can
see at the moment would be if there was another extension module
involved that was expecting to be called back on the same thread ...