Reorganise and comment ext.c
The code in ext.c involves multiple contexts and jumps between them.
This is confusing enough without leaving the functions in the
haphazard ordering they were in.  Tidy things up by grouping functions
that run in the same context, and also add comments so that it's
easier to see what's going on.