Simplify ext.c code a bit
The way that g0_main was resetting the state on entry meant that the
only way we could get in with something to do was to jump in using
setcontext - this made the code more complicated than necessary.
Refactor things so that we can get g0_main to do something useful on
the first time through if we want.

Also add a few more comments here and there, and tidy a few things up
a bit.