Make sure that the runtime knows about indirections
We want to keep these indirections, which means that the Go runtime
needs to know that we are holding references to them.  It won't look
at the contexts structs, because they are C types and so opaque to the
runtime - so we need to store them separately in a way the runtime
will see them.