Start work on modernising
This is an attempt to bring things up to date, both in terms of Go and
in terms of Python.

At the moment the code is unchanged, but has been moved around a bit and
the python import changed to 3.11 - with the intention of targeting that
version instead of the ancient 2.7 that was being targeted.