Start working on module importing
Python has an all new import system that is almost entirely implemented
in Python now. There is a mechanism to hook in modules implemented in C,
but unfortunately it requires the use of a C function that takes no
arguments per module.

There is no way to manage that, without passing on the requirement to
create a C function to library users. Which is a deal breaker really, I
absolutely do not want the library users to have to mess around with cgo
stuff.

So instead of trying to use the C hooks, start looking at implementing a
custom importer instead. The importer will directly load the Go modules,
without requiring hooking into the BuiltinLoader that loads the C
modules. This requires implement a Python Object that has the requiste
methods to perform the loading. The default Python loaders, such as
BuiltinLoader, use classes with static methods.  This seems like a good
idea, only the class support doesn't currently allow for static or class
methods.