Start trying to create an importer class
Start trying to create an importer class that we can insert into
sys.meta_path to hook into the module loading mechanism.

This also means that we need to start making some tweaks to the class
support as we want to create a class that has only static methods, and
which doesn't have a custom object type associated with it.