Start changing the way that we do classes
Previously we used to play some rather nasty games to get Go managed
memory used for holding Python objects. Go has since tightened the rules
for how memory can be used to interface with cgo, and these tricks no
longer work.

Instead of trying to get all clever with memory tricks, we now take a
different tack. We use a proxy object. The Go object never gets used in
Python, instead we allocate a basic PyObject in Python and use that, and
we have a lookup table to retrieve the Go version when any of the
methods are called.

Currently the transition is not complete, so there are still going to be
functions that are trying to convert the python object directly into a
Go object, but I think the core should be in place now.