Some more essential changes for the new approach
newObject needed updated to use the new mapping facitily to conver the
Python proxy object into the appropriate Go object.

In addition, we can't use Type.Alloc as that returns a mapped Go Object,
which we don't want in this case. Since we are allocating memory for the
Python proxy object we don't want it to be mapped to a Go Object.