Add support for class methods to get importer working
The find_spec method is supposed to pass itself to spec_from_loader as
the loader, not the module that we found. Since we can refer to importer
directly in importerFindSpec since importer already refers to the
function, use a class method instead. Using a class method solves the
issue as it means that the Class (i.e. importer) gets passed in as the
first argument.