Move more class functions to new mechanism
A lot of the class functions are still using the old mechanism to
translate between Go and C, and this means that they aren't working
correctly. Start updating them to use the new ClassObject based
approach.

In the case of the slot functions, we can actually also get rid of the
horriby function pointer hacks, and just use interfaces instead to get
direct access to the appropriate methods.