Create some constants for new context setup
Since we are switching to using interfaces for the slot methods, we
don't need to store pointers in C (which is nice). So we only need one
bit for each slot - to indicate if we should set it or not. So start by
mapping slots to flags for a uint64_t flags value.