Start figuring out which slot flags apply
We need to convert all of the call functions over to using interfaces
before we can complete this part. Since we need the interfaces that go
with that before we can fully populate slotMap. Not to mention that we
want to convert all the call methods anyway, as well as removing the
functions from the context struct.