Change the public API a bit
abstractObject and the xxxProtocol types shouldn't be public. These
types are part of the internal implementation and have no place in the
public API. They don't make sense to be used outside of this package,
and they can't be used when implementing a new type via Class, so the
only benefit is that they show the types that implement the interfaces.