Start expanding the Chan API
Add some more more methods to the Chan type. Including implementing
iterator methods so that we can use "for o in c:" in Python to iterate
until the channel is closed. Also add a close method.

As part of this I also realised that abstractObject wasn't working
properly. If the pointer was nil then it was causing a nil dereference
when calling a method on the astractObject. Since we are now generating
the bulk of the types, we can simply include the abstractObject methods
in the generated code and have them as direct methods, which fixes the
issue. It also has the advantage that the methods will show up in the
documentation.

I also switched the example to use InitAndLockWithSignals, and that
showed up an issue with the lock code that was causing a SIGSEGV - this
is now a more friendly Go panic that at least tries to explain the
issue.