Raise exception when channel is closed
When the Go channel is closed, then have the Chan methods raise a
ChanClosedError exception. This allows the Python code to deal with the
channel being closed in an idiomatic Python manner.

This also meant adding the function to create exceptions in Go, and
adding the new exception to the Go module so that it can be accessed
from Python.