Add support for sub-modules
Update the importer and module code to support sub-modules. We can now
mark a module as a package, which will add the __path__ member needed
for the import machinery to recognise it as a package. The importer will
also import modules that are sub-modules, provided that we also provide
the parent package.

The name of the sub-module should be "<parent>.<module>".