Improve reference handling for GetIterator
If GetIterator returns the object given, then increment the reference
count. If we don't do this then the caller can't manage the reference
count properly, as calling Iter() will return a new reference. This way
the caller knows that they have to decrement the count unless an error
is returned.