Sort out traverse and clear for classes
Implement the traverse and clear functions using the new approach to
class objects.

Also fix some issues with dealloc - we should be clearing out any
contained references here too. So now we do so using the same code as in
the clear slot, unless there is a dealloc method defined. In which case
that method is responsible, though it can call py.ClearClassObject to
acheive the same thing.