Reduce the Object interface
I don't think it makes sense to have IsTrue, Not or Free in the Object
interface. These are not common methods to be calling, and I think it is
fine to only have them on BaseObject, so the more general case is
o.Base().IsTrue() instead of o.IsTrue().

Also add a raw method to objects that does the same thing as the c()
function, but without having to worry about a nil interface. The idea
being that we can use o.raw() when we have a concrete type and avoid the
need to create an interface and then test it for nil. Though that
conversion to using raw methods has not been done yet.