Introduce RefManager
The idea of the RefManager is to try and help with reference management.
You can setup the RefManager with a defered decref at the top of a
function and then add and remove items as you go.

Unfortunately I don't think the E versions of the methods are going to
be as useful as I hoped without the ability to have generic methods (so
that they can return the exact type of the argument instead of
converting it to Object).