Add Support for Super
In Python subclasses can use super to call methods of the base class.
This is actually a type implemented in C as PySuper_Type. So introduce a
Super equivalent, and add a Super method to Class so that user
implemented classes can use it to access methods of the parent class.

Also fix a typo in the gen_types template that I spotted when adding
Super.