Start reworking to support subclassing Classes
In order to support being able to subclass a Class we need to change the
mechanisms of how we discover the class object to allow for multiple Go
objects against a single Python object.

Currently this is only partially complete. The generated slots code has
been updated, but the other parts such as member access still need to be
updated.