Update methods to include type in context
Update the method wrapper to include the type that we are registering
for, so that we can make sure we lookup the correct type even when the
current type is a subtype.

I think that class methods will not work properly if the subclass is
Python as they will not be able to find a Class to pass as the first
argument. Not sure what to do about this, passing the raw Type instead
doesn't seem very useful.