Add AsXXX methods to allow more casting
Since we allow BaseType to be a *Type, that means that a user may want
to cast to a built-in type like Dict from a Class instance like MyDict.
So add AsXXX methods to support this. So now MyDict can use AsDict to
convert the instance pointer to a *Dict to access the Dict method set.