Sequences are always iterable so expose that
There is a dedicaed iterator than will turn any sequence into an
iterator. So add an Iter method to SequenceMethods - use
PyObject_GetIter rather than PySeqIter_New directly though, so that
dedicated iterators are used in preference.

Also check if a type is a sequence in AsIterable, since any sequence
should be returned, and we can use AsSequenceMethods for that.