Start adding support for heap allocated Class
The first thing is that we can use PyHeapTypeObject to replace the use
of ClassContext to store the method sets. Doing this regardless of the
heap flag makes things simpler in addition to aligning with the layout
we will need for a heap allocated Class. We also switch how we allocate
memory based on the flag, for heap allocated we use the normal object
allocation functions, so that Python recognises the type as an object.

There is still more work to do before we will behave correctly when
setting the flag though.