Add Chan.monitor to run callback in the background
The new monitor method starts a goroutine that calls the provided item
callback for each item received from the channel. There is also an
optional closed callback that will be called when the channel is closed.

Since this method uses a background goroutine it needs the GIL to be
released in order to process items from the channel.