Address a bunch of linter errors
Since the code was written without linting, there are quite a few
complaints to address. This is only some of them.

In some cases the linter config is tweaked instead of changing the code.

The single non-stdlib dependency is removed, since the Ordered type used
in generics is now available from the cmp package in the standard
library.