Refactor goClassNatSet for readability
This was a long function, making it hard to read, so pull a load of code
out into helper functions to make it more readable.

Also add some panic recovery code so that in the unlikely case that a
panic occurs (since we are calling methods that are documented as using
panic on failure) we convert it into a Python exception instead of
letting the panic run back into the Python code.